+++ pages/views/utils/esl-base-element.njk
@@ -1,7 +1,7 @@
 ---
 layout: content
 title: ESL Base Element
-seoTitle: ESL Smart Library Base custom element definition
+seoTitle: ESL Smart Library base custom element definition
 name: ESL Base Element
 tags: utils
 iconAlt: '<esl-.../>'
