+++ pages/views/utils/esl-mixin-element.njk
@@ -1,0 +1,11 @@
+---
+layout: content
+title: ESL Mixin Element
+seoTitle: ESL Smart Library base mixin definition
+name: ESL Mixin Element
+tags: [utils, beta]
+iconAlt: '<div esl-.../>'
+---
+
+{% mdRender 'src/modules/esl-mixin-element/README.md', 'intro' %}
+
