+++ src/modules/all.ts
@@ -1,29 +1,41 @@
+// Core
+export {ESLBaseElement, ESLBaseElementShape} from './esl-base-element/core';
+export {ESLMixinElement} from './esl-mixin-element/core';
 export * from './esl-utils/all';
-
 export * from './esl-media-query/core';
 export * from './esl-traversing-query/core';
+
+// Helpers
 export * from './esl-a11y-group/core';
 
+// Image
 export * from './esl-image/core';
 
+// Media
 export * from './esl-media/core';
 
-export * from './esl-toggleable/core';
-export * from './esl-popup/core';
+// Toggleables
 export * from './esl-trigger/core';
-
-export * from './esl-panel-group/core';
-export * from './esl-panel/core';
+export * from './esl-toggleable/core';
 
 export * from './esl-tab/core';
+export * from './esl-panel/core';
+export * from './esl-panel-group/core';
 
+export * from './esl-popup/core';
+
+// Scrollbar
 export * from './esl-scrollbar/core';
 
+// Alert
 export * from './esl-alert/core';
 
+// Forms (beta)
 export * from './esl-forms/all';
 
+// Footnotes
 export * from './esl-footnotes/core';
 export * from './esl-tooltip/core';
 
+// Animate
 export * from './esl-animate/core';
