+++ src/modules/esl-base-element/README.md
@@ -8,9 +8,7 @@ Authors: *Alexey Stsefanovich (ala'n)*
 
 Provides the core for creating custom elements and ts decorators to simplify components creation.
 
-## Base Element
 **ESLBaseElement** - base class for custom (tag) element declaration
-**ESLMixinElement** (beta) - base class to create a mixin element that attaches to the element via a custom attribute
 
 ### Base Element static API
 - `MyElement.is` - property that defines tag name
@@ -18,13 +16,6 @@ Provides the core for creating custom elements and ts decorators to simplify com
 
 - `MyElement.register` - register component inside `customElements` registry
 - `MyElement.registered` - returns promise that will be resolved as soon as the component is registered
-
-### Mixin Element static API (beta)
-- `MyMixinElement.is` - property that defines connection attribute name
-- `MyMixinElement.observedAttributes` - array of additional attributes to observe
-
-- `MyMixinElement.register` - register component inside `ESLMixinRegistry`
-
 
 ### Base Element API
 Properties:
@@ -42,11 +33,7 @@ Attributes:
 - `$$on` - subscribe on event manually or subscribe decorated method
 - `$$off` - unsubscribe from event manually or unsubscribe decorated method
 
-### Mixin Element API (beta)
-- `$host` - readonly mixin target DOM element
-
 ### Element decorators
-Works for both `ESLBaseElement` and `ESLMixinElement`.
 
  - `@attr` - to map string type property to HTML attribute
  - `@boolAttr` - to map boolean property to HTML boolean (marker) attribute state
@@ -143,37 +130,4 @@ class MyCustomComponent {
 
   onEvent(e: Event) { /* ... */ }
 }
-```
-
-### Mixin Example (beta)
-
-```ts
-import {ESLMixinElement, attr, boolAttr, jsonAttr, listen} from '@exadel/esl';
-
-class MyMixinComponent extends ESLMixinElement {
-  static is = 'my-mixin-attr';
-
-  /** Reflects 'my-string-prop' attribute */
-  @attr() public myStringProp: string;
-  /** Reflects to 'my-marker' attribute-marker */
-  @boolAttr() public myMarker: boolean;
-  /** Reflects to JSON value in 'my-config' attribute */
-  @jsonAttr() public myConfig: Recorg<string, string>;
-
-  connectedCallback() {
-    super.connectedCallback();
-    // Init my component
-  }
-
-  disconnectedCallback() {
-    // Unsubscribe listeners, revert side effects
-    super.disconnectedCallback();
-  }
-
-  @listen('click')
-  onClick(e: MouseEvent) { /* Handle btn click event */ }
-}
-
-// Register mixin element for attribute provided in the static `is` property
-MyMixinComponent.register();
 ```
