+++ src/modules/esl-mixin-element/core.ts
@@ -1,6 +1,4 @@
-export type {ESLBaseElementShape} from './core/esl-base-element.shape';
-
-export * from './core/esl-base-element';
+export * from './ui/esl-mixin-element';
 
 export * from '../esl-utils/decorators/attr';
 export * from '../esl-utils/decorators/bool-attr';
